Fix match positions by patching Annot.TextBuilder
For example, in the Markdown document

"Test Document
=============

This is a `code *test* boing`.
"

the unknown word "boing" wouldn't be underlined exactly, but
rather the characters "oing`". This is due to AnnotatedText using
a Map to save the associations between original text and checked
text. Everytime addMarkup(String markup) was called, the current
association was overwritten in the Map, as Maps can only have
a single value for each key. This is fixed by using a list of
pairs of integers instead. Unfortunately, this means that
AnnotatedTextBuilder has to be monkey-patched as well.